+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+    - rtl (right to left) support
+
 # v0.2.4
     - resize embedded images