+++ CHANGES.md
@@ -1,8 +1,7 @@
 
-    - rtl (right to left) support automatically right-aligned
-
-# v0.2.4
+# v0.2.4 2024.05.24
     - resize embedded images
+    - rtl (right to left) support automatically right-aligned
 
 # v0.2.3 2024.05.16
     - Add the `embed` field to embed an image.